Refactor Tabs implementation and update routing in TabsPage component
