+++ src/Root.tsx
@@ -8,11 +8,10 @@ export const Root = () => (
     <Routes>
       <Route path="/" element={<App />}>
         <Route index element={<HomePage />} />
-        <Route path="/home" element={<Navigate to={'/'} />} />
-        {/* <Route path="/tabs" element={<TabsPage />} /> */}
+        <Route path="home" element={<Navigate to={'/'} />} />
         <Route path="tabs">
           <Route index element={<TabsPage />} />
-          <Route path=":tabId?" element={<TabsPage />} />
+          <Route path=":tabId" element={<TabsPage />} />
         </Route>
         <Route path="*" element={<h1 className="title">Page not found</h1>} />
       </Route>
